% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_uind.R
\name{add_uind}
\alias{add_uind}
\title{Add Independent Unif Predictors to Data}
\usage{
add_uind(dat, rv, ord = NULL)
}
\arguments{
\item{dat}{Data, as in the output of \code{\link{inputdat}}, containing
the \code{$u} output. Should also have the \code{$ycol} output,
otherwise will assume that the predictors have labels 1,2,...,p, where p
is the number of predictors.}

\item{ord}{Order of predictors to compute.}

\item{cnqr_obj}{Object of type \code{'rvine'}, describing at least the
predictors in \code{dat}.}
}
\value{
Returns the original \code{dat}, but with the \code{$uind}
entry filled out.
}
\description{
Adds the independent uniform predictors "\code{uind}" to specified data,
by computation via \code{\link{pcondseq}} in the \code{copsupp} package.
Does nothing if \code{uind} is already present in the data.
}
\details{
Note: This function, along with \code{\link{inputdat}}, will probably
become deprecated.
}
\note{
This is slightly different from \code{\link{pcondseq}}, which
computes the PIT scores of *each variable* given previous ones in a
provided vine.
}
