% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adddat.cnqr.R
\name{adddat.cnqr}
\alias{adddat.cnqr}
\title{Add data to cnqr object}
\usage{
\method{adddat}{cnqr}(obj, dat, QY = NULL)
}
\arguments{
\item{obj}{Object of type \code{'cnqr'} that you wish to add
data to. See \code{\link{cnqr}}.}

\item{dat}{A list of data frames/matrices (of the raw data) to add
to \code{obj}. Or, a single data frame/matrix if there's only one.}

\item{QY}{List of quantile functions of length matching the number
of data matrices in \code{dat} (or a single quantile function if there's
only one data matrix). \code{NULL} if you'd like to use the quantile
function in \code{obj} -- in which case the response data are stationary.}
}
\value{
The original \code{'cnqr'} object \code{obj}, with the
\code{$y}, \code{$uind}, \code{$yhat}, \code{$score}, and possibly
\code{$QY} (if non-stationary) entries extended to include the new data.
}
\description{
Add more data to a \code{'cnqr'} object.
}
\note{
If \code{dat} is a named list (and not a matrix), then the data added to the
\code{'cnqr'} object will also have those names. Otherwise,
In the outputted \code{'cnqr'} object, the added data will have names
\code{'dat'} with the data number beside it. So for example,
if \code{obj} contains two data sets, and you want to add two more unnamed
data sets, they'll be named \code{'dat3'} and \code{'dat4'}.

This function is useful for cutting down on
computation time. Particularly, it helps so that the independent
predictors are only needed to be computed once.
}
\examples{
data(egdat)
dat <- list(egdat[1:750, ], egdat[751:1000, ])
basevine <- subset(egvine, 1:4)
sc <- scorer(space_taus(10))
fit <- cnqr(5:3, dat, sc, basevine=basevine,
            families=c("bvncop", "joe"))

set.seed(123)
newdat <- rrvine(100, egvine)
summary(adddat(obj, dat=newdat))
summary(adddat(obj, dat=list(test=newdat)))
}
