% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymloss.R
\name{asymloss}
\alias{asymloss}
\alias{rho}
\title{Asymmetric absolute deviation function}
\usage{
asymloss(tau, u)

rho(tau, u)
}
\arguments{
\item{tau}{Number in [0,1] representing the probability/order
of the quantile.}

\item{u}{Value to evaluate the loss function at. Vectorised.}
}
\description{
The loss function that defines the tau^th-quantile;
\code{asymloss(tau, u) = (tau - I(u<0))*u}. Vectorised over \code{u}.
Function \code{rho()} is deprecated; use \code{asymloss()} instead
(they're identical).
}
