% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpar_init.R
\name{cpar_init}
\alias{cpar_init}
\title{Get Initial Values for Copula}
\usage{
cpar_init(u, v, cop)
}
\arguments{
\item{u, v}{Vectors of uniform values.}

\item{cop}{Character of a (single) copula family to consider fitting.}
}
\description{
Obtains initial parameter estimates for a bivariate copula, using MLE.
This includes the copula reflection. Uses \code{\link{fitbicop_lh}}.
If the IG, IGL, or skew normal copula families are requested, then MLE
is not used (since the copulas are not recognized by the VineCopula pkg).
Pre-chosen values are output instead, and no reflections are considered.
}
