% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egdat_documentation.R
\docType{data}
\name{egdat}
\alias{egdat}
\alias{egvine}
\title{Sample data from a Vine}
\format{\code{egdat} is a matrix with 1000 rows and 5 columns of
uniform variables.

\code{egvine} is an object of type \code{'rvine'} describing the joint
distribution of \code{egdat}.}
\usage{
egdat

egvine
}
\description{
\code{egdat} is a randomly generated dataset of 5 variables described by the
vine \code{egvine}, using the seed 555 and the function \code{\link{rrvine}}.
The vine is a D-Vine.
}
\keyword{datasets}
