% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggally_qsmooth.R
\name{ggally_qsmooth}
\alias{ggally_qsmooth}
\title{Plots the Scatter Plot with Quantile Smoothing}
\usage{
ggally_qsmooth(data, mapping, ..., method = "lp", tau = 1:9/10,
  npoints = 100, h = 1, kernel = "gaussian")
}
\arguments{
\item{data}{Data frame of data}

\item{mapping}{Aesthetics to use}

\item{...}{Other arguments to pass to \code{geom_point} for the underlying
scatterplot.}

\item{method}{The smoothing method to use. Either
\code{"lp"} (default) for local polynomial regression with \code{\link{lpqr}},
\code{"ss"} for spline regression with \code{quantreg}'s \code{rqss()}, or
\code{"gpd"} for kernel smoothing (i.e. local polynomial with degree 0)
with a GPD tail.}

\item{tau}{Vector of quantile indices to plot curves for.}

\item{npoints}{Number of points to evaluate each quantile curve at.}

\item{h}{Numeric smoothing parameter, passed to the \code{h} argument
in the function \code{\link{lpqr}} if \code{method="lp"}, or
passed to the \code{lambda}
argument in the function \code{qss()} in package \code{quantreg} if
\code{method="ss"}. For specifying a different smoothing parameter
depending on which variable is used as a covariate, make this a list
of such parameters -- see details for how to do this.}

\item{kernel}{Type of kernel to use. Passed to the argument \code{kernel}
in \code{\link{lpqr}}.}
}
\description{
Uses splines to fit smoothed conditional quantile curves overtop of a
scatter plot. Intended for use with \code{ggpairs} in package
\code{GGally}.
}
\details{
The quantile curves are estimated using the \code{rqss} function
in package \code{quantreg}. For response \code{y} and covariate \code{x},
the function call for a single \code{tau} is
\code{rqss(y ~ qss(x, lambda=h), tau=tau, data=data)}.

There are two ways of specifying a list for \code{h}:

\enumerate{
     \item A named list of smoothing parameters with names
     in \code{names(data)}, and possibly a component named \code{".default"}.
     The smoothing parameter will be selected according to the name
     of the \code{X} variable, or will use the \code{.default} if
     the name is not present in \code{h}.
     \item An unnamed list of length equal to \code{ncol(data)}. The
     smoothing parameter will be selected according to the column number
     of the \code{X} variable.
}

This function should be used when calling \code{ggpairs} in package
\code{GGally}, under the "continuous" option of argument \code{upper}
and \code{lower} (for which you can specify the string \code{'qsmooth'}).
See examples.
}
\note{
The curves are fit to the entire scatterplot, even if groups are
somehow specified (such as through the \code{colour} aesthetic).
}
\examples{
library(GGally)
library(ggplot2)
set.seed(123)
x <- runif(1000)*pi
y <- abs(rnorm(1000))*sin(x)*100
z <- x + rexp(1000)*y
dat <- data.frame(x=x, y=y, z=z, grp=factor(x<pi/2))

## Plain function call
ggpairs(dat, columns=1:3, lower=list(
  continuous="qsmooth"
))

## Increase the smoothing parameter
ggpairs(dat, columns=1:3, lower=list(
  continuous=wrap("qsmooth", h=5)
))

## Increase the smoothing parameter, just for Y:
ggpairs(dat, columns=1:3, lower=list(
  continuous=wrap("qsmooth", h=list(.default=1, y=5))
))

## Add other specifications
ggpairs(dat, columns=1:3, lower=list(
  continuous=wrap("qsmooth", h=list(.default=1, y=50),
                  tau=16:19/20,
                  alpha=0.25,
                  method="ss"),
  mapping=aes(colour=grp)
))
}
\references{
The code for this function is modelled in part after
\code{ggally_smooth} in package \code{GGally}.
}
