% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdm.R
\name{pgpdm}
\alias{pgpdm}
\alias{qgpdm}
\title{Mixture Generalized Pareto Distribution}
\usage{
pgpdm(x, mu = 0, beta = 1, xi = 0, basecdf)

qgpdm(p, mu = 0, beta = 1, xi = 0, baseqf)
}
\arguments{
\item{x}{Vector of values to evaluate the distribution at.}

\item{mu, beta, xi}{Numeric; parameters of the GPD function (truncation,
scale, and extreme value index, respectively).}

\item{basecdf, baseqf}{Vectorized distribution function/quantile function
to use below the truncation \code{mu}.}
}
\value{
Vector of evaluated distribution function, or quantiles.
}
\description{
Distribution and Quantile functions for a mixture Generalized
Pareto Distribution. In particular, the tail of a "base" distribution
(above some threshold) is overridden with a GPD.
}
\examples{
cdf <- function(y) pgpdm(y, mu=2, xi=2, basecdf=pnorm)
qf <- function(tau) qgpdm(tau, mu=-2, xi=0, baseqf=qnorm)
curve(cdf, 0, 5)
curve(qf, 0, 1)

set.seed(123)
x <- rnorm(100)
basecdf <- ecdf(x)
baseqf <- function(tau) quantile(x, tau)
cdf <- function(y) pgpdm(y, mu=0, beta=5, xi=-2, basecdf=basecdf)
qf <- function(tau) qgpdm(tau, mu=-1, xi=-1, baseqf=baseqf)
curve(cdf, -3, 5, )
curve(qf, 0, 1)
}
