% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputdat.R
\name{inputdat}
\alias{inputdat}
\title{Input Data}
\usage{
inputdat(xy = NULL, xv = NULL, uy = NULL, uv = NULL,
  uindy = NULL, uindv = NULL, yx = NULL, vx = NULL, yu = NULL,
  vu = NULL, yuind = NULL, vuind = NULL, y = NULL, v = NULL,
  x = NULL, u = NULL, uind = NULL, prevdat = NULL, ycol = NULL)
}
\arguments{
\item{xy}{Data frame or matrix of \code{x} and \code{y} data.
If \code{ycol} is \code{NULL}, response is in the last column.}

\item{xv}{Data frame or matrix of \code{x} and \code{v} data.
If \code{ycol} is \code{NULL}, response is in the last column.}

\item{uy}{Data frame or matrix of \code{u} and \code{y} data.
If \code{ycol} is \code{NULL}, response is in the last column.}

\item{uv}{Data frame or matrix of \code{u} and \code{v} data.
If \code{ycol} is \code{NULL}, response is in the last column.}

\item{uindy}{Data frame or matrix of \code{uind} and \code{y} data.
If \code{ycol} is \code{NULL}, response is in the last column.}

\item{uindv}{Data frame or matrix of \code{uind} and \code{v} data.
If \code{ycol} is \code{NULL}, response is in the last column.}

\item{yx}{Data frame or matrix of \code{y} and \code{x} data.
If \code{ycol} is \code{NULL}, response is in the first column.}

\item{vx}{Data frame or matrix of \code{v} and \code{x} data.
If \code{ycol} is \code{NULL}, response is in the first column.}

\item{yu}{Data frame or matrix of \code{y} and \code{u} data.
If \code{ycol} is \code{NULL}, response is in the first column.}

\item{vu}{Data frame or matrix of \code{v} and \code{u} data.
If \code{ycol} is \code{NULL}, response is in the first column.}

\item{yuind}{Data frame or matrix of \code{y} and \code{uind} data.
If \code{ycol} is \code{NULL}, response is in the first column.}

\item{vuind}{Data frame or matrix of \code{v} and \code{uind} data.
If \code{ycol} is \code{NULL}, response is in the first column.}

\item{y}{Vector of the response on its original scale.}

\item{v}{Vector of the PIT response}

\item{x}{Matrix of the predictors on their original scale.}

\item{u}{Matrix of the PIT predictors.}

\item{uind}{Matrix of the independent uniform predictors (for example,
the output of \code{\link{pcondseq}}).}

\item{prevdat}{Output of previous \code{inputdat}.}

\item{ycol}{The column number of the response in the input data frame.}
}
\value{
A list containing as many of the following entries as can be filled:

\itemize{
   \item \code{$y}: Vector of the response variable.
   \item \code{$v}: Vector of the PIT response variable.
   \item \code{$x}: Matrix of predictors.
   \item \code{$u}: Matrix of PIT predictors.
   \item \code{$uind}: Matrix of independent-transformed PIT predictors.
   \item \code{$ycol}: Integer; "variable number" of the response, corresponding
   to its column number in the original data matrix.
}

If an entry can't be filled, then the entry is \code{NULL}.
For predictors \code{x}, \code{u}, and \code{uind}, columns are in the same
order as the original input.

\code{ycol} is useful to include in the output, so that computations on the
predictors correspond to the appropriate columns. For example,
\code{\link{pcondseq}} thinks that the inputted data matrix has columns
that match their variable numbers -- but this may not be the case if
computing on the predictor matrix only, *which is extracted* from the
larger data that includes the response.
}
\description{
Easy specification of data for CNQR, by indicating predictors and
response in an intuitive way. Don't get scared off by the number
of arguments: their abundance allows flexibility of input. See details
for how to use this function; or more tediously, you can read the
argument descriptions.
}
\details{
NOTE: This function may become deprecated. I thought it would be convenient
to work with one data object containing the various transformations of the
data, but all that's needed is the uniform scores data anyway.

Data can either be in their original form
(predictors \code{x} and response \code{y}), or
in their probability integral transform (PIT) scores
(predictors \code{u} and response \code{v}). Additionally, the predictors
can be independent uniform predictors (\code{uind}).

You have three options for inputting whichever of these data you have.

\enumerate{
   \item Indicate them directly, by the arguments \code{y}, \code{v},
    \code{x}, \code{u}, and/or \code{uind}.
   \item Indicate them as a data frame or matrix, by the arguments
    \code{<predictor><response>} (response in last column) or
    \code{<response><predictor>} (response in first column), where
    \code{<predictor>} is one of \code{x}, \code{u}, or \code{uind}, and
    \code{<response>} is one of \code{y} or \code{v}.
   \item Indicate them as in 2., but override the position of the response
    by using the \code{ycol} argument.
}

Whenever data are in a matrix or data frame, columns should represent
variables, and rows observations. Predictors should be input as matrices
or data frames.

If you wish to fill more of the output list from a previous \code{inputdat}
output, specify this previous data in the \code{prevdat} argument.
}
\note{
This function doesn't check for errors, such as unmatching
number of observations.
An error won't be thrown if you try to input data in more than one
way, nor if those data don't match up. The final data to be used
is taken in the order that the arguments appear, with later arguments
overriding previous.
}
\examples{
inputdat(y=rnorm(10))

(dat <- inputdat(uv=matrix(runif(20), ncol=2)))
inputdat(x=rnorm(dat$u), y=rnorm(dat$v), prevdat=dat)

inputdat(xy=matrix(rnorm(10)))
}
