% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpqr.R
\name{lpqr}
\alias{lpqr}
\title{Initiate a Local Polynomial Quantile Smoothing}
\usage{
lpqr(formula, data, h = 0.5, kernel = "gaussian")
}
\arguments{
\item{formula}{A formula indicating the response, and the covariates
to smooth over. Should be of the form "response ~ covariate_1 + ... +
covariate_p" (higher order and interaction terms are meaningless here).}

\item{data}{Data frame containing the data used for smoothing.}

\item{h}{Bandwidth(s). If a single scalar, specify in
units of mahalanobis distance.
If a vector (of length equal to the number of covariates, for 2 or more
covariates), each covariate is scaled by the bandwidths in \code{h}
(which should be specified in the same order as the \code{formula}).}

\item{kernel}{Character; type of kernel to use. One of
\code{"gaussian"}, or \code{"boxcar"}. See details for description.}
}
\value{
Object of class \code{"lpqr"}, which is a list of the
following components.

\describe{
     \item{extractx}{Function that takes a data frame containing
     at least columns of the covariates, and returns a
     data frame of the covariates indicated in \code{formula}.}
     \item{center}{Function that takes a covariate vector \code{x0} and
     returns a data frame of the response and covariates centered
     at \code{x0}.}
     \item{w}{Function that takes a covariate vector \code{x0} and
     returns a vector of normalized weights used for local polynomial
     regression, to be applied to the observations in \code{data}.}
     \item{data}{The original \code{data}.}
     \item{xnames}{Character vector of covariate names.}
     \item{yname}{Character name of the response.}
}
}
\description{
Sets up a local polynomial quantile regression smoother.
}
\details{
Since local polynomial regression needs to be done "on-the-spot",
the regression isn't actually done here. Instead, this function
sets up the regression by primarily building a weight function and a
data-centering function.

Here are the kernel functions for distance \code{d} and bandwidth \code{h}:

\describe{
     \item{"gaussian"}{\deqn{K(d,h) = \exp (-(d/h)^2/2)}}
     \item{"boxcar"}{\deqn{K(d,h) = I(d \le h)}}
}
}
\note{
There is a similarly named function in the package \code{quantreg},
called \code{lprq()}, that only allows for one covariate. \code{lpqr()}
is intended to extend this, using S3 generic functions. The nomenclature
of \code{"qr"} is used to follow that of \code{cnqr()}.
}
\examples{
## See predict.lpqr() for examples.
}
\seealso{
\code{\link{predict.lpqr}}
}
