% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.cnqr.R
\name{predict.cnqr}
\alias{predict.cnqr}
\title{Make Quantile Predictions}
\usage{
\method{predict}{cnqr}(obj, dat = "tr", tau = NULL, QY = NULL)
}
\arguments{
\item{obj}{Object of type \code{'cnqr'}. See \code{\link{cnqr}}.}

\item{dat}{Data matrix or data frame with raw data.
Or, a character vector specifying data that's already in \code{'obj'} --
for example, \code{"tr"} for the training data (always present), or
\code{"val"} for validation data. The data specified will be pooled.}

\item{tau}{Vector of quantile levels to predict. If \code{NULL}, uses
the quantile levels used in the regression of \code{obj}.}

\item{QY}{(Quantile) function that accepts a vector of quantile levels, and either:
(1) returns an equally long vector of corresponding quantiles, or
(2) returns a matrix of quantiles, with columns corresponding to the
quantile level input, and rows corresponding to the observations/rows in
\code{dat}. Alternatively, \code{NULL} if the quantile function is
already in \code{obj} (and you specify \code{dat} as a character).}
}
\value{
A matrix of quantiles, with rows corresponding to observations, and columns
corresponding to quantile levels.
}
\description{
Use a \code{'cnqr'} object to make quantile predictions of a response
given predictors.
}
\details{
The response is taken to be the upper-right variable in the vine array of
\code{obj}, and the predictors are the variables below the response.
The vine is used to make quantile predictions of the conditional response.
}
\examples{
data(egdat)
dat <- list(egdat[1:750, ], egdat[751:1000, ])
basevine <- subset(egvine, 1:4)
sc <- scorer(space_taus(10))

fit <- cnqr(5:1, dat, basevine, sc=sc)

head(predict(fit))
head(predict(fit, tau=c(0.1, 0.5, 0.9)))

set.seed(442)
newdat <- rrvine(10, egvine)
predict(fit, dat=newdat)
}
