% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpqr.R
\name{predict.lpqr}
\alias{predict.lpqr}
\title{Evaluate Local Polynomial Quantile Surfaces}
\usage{
\method{predict}{lpqr}(object, newdata = object$data, p = 1,
  tau = 1:9/10)
}
\arguments{
\item{object}{Object of class \code{"lpqr"} (see \code{\link{lpqr}}).}

\item{newdata}{A data frame containing at least covariate columns for which
to evaluate the regression surface(s) at.}

\item{p}{Non-negative integer; power of local polynomial
(which includes all interaction terms of that order). If you'd like more
flexibility, you can also specify
a formula here for the local regression.}

\item{tau}{Quantile indices indicating which quantile surfaces to evaluate.}
}
\value{
A matrix, with columns corresponding to the quantile
indices \code{tau}, and columns corresponding to the query points in
\code{newdata}.
}
\description{
Performs local polynomial quantile regression by evaluating the regression
surface(s) at query points (or "new data").
}
