% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_forecaster.R
\name{score_forecaster}
\alias{score_forecaster}
\title{Score a Forecaster}
\usage{
score_forecaster(forecaster, y, x, K = 10, tau_c = 0.9, g = identity,
  wtau = function(tau) 1/tau/(1 - tau), wx = 1, cmb = TRUE,
  yhat = NULL)
}
\arguments{
\item{forecaster}{Function with first argument being the quantile index
\code{tau}, which
should accept a vector. Optionally, accepts a matrix of predictors
\code{x} as a
second argument. Returns either a vector of quantiles of length equal to
the length of \code{tau} if predictors are not accepted, or a matrix
of quantiles with columns corresponding to \code{tau}, and rows
corresponding to rows of \code{x}.}

\item{y}{Response vector}

\item{x}{Matrix of predictors. Optional if the forecaster doesn't
use predictors.}

\item{K}{Number of points to use when estimating the integral.}

\item{tau_c}{Lower quantile index of the forecasts.}

\item{g}{Vectorized univariate mapping function of the response and forecasts.
See formula in details.}

\item{wtau}{Function that accepts a vector of quantile indices and returns
a weight when assessing that quantile. See formula in details.}

\item{wx}{Function that accepts a vector of predictors (so, a row of \code{x})
and returns a weight value to adjust that observation's score by. Or,
a vector of weights already evaluated (no need to normalize them).}

\item{cmb}{Combine the scores across quantiles and observations?
\code{TRUE} by default. (See "Value" section for details of what's output)}

\item{yhat}{Optional matrix of predictions. It could be a time saver
because the forecasts don't need to be computed. However, be sure that
the quantiles that are forecast match up with \code{K} evenly spaced
quantile indices in \code{(tau_c, 1)}.}
}
\value{
If \code{cmb=TRUE} (default), the mean score of the forecaster
is returned (i.e. a single numeric). Otherwise, a matrix of
(weight-adjusted) scores for each
observation (rows) and quantile index (columns) is returned, with columns
labelled by the quantile indices (rounded to four decimals).
}
\description{
Score a Forecaster
}
\details{
If y_i is the i'th observation,
tau_k = (2k-1)/(2K)*(1-tau_c) + tau_c is the k'th quantile index,
yhat_{ik} is the tau_k-quantile forecast for observation i, and
wx'_i = n wx_i / sum wx_i where wx_i = wx(x_i), then the scores
are computed as
S_{ik} = wx'_i wtau(tau_k) rho_{tau_k}(g(y) - g(yhat_{ik})).
These are returned in a matrix if \code{cmb=FALSE}, and are averaged
otherwise.
}
