% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer.R
\name{scorer}
\alias{scorer}
\title{Define a Scoring Rule}
\usage{
scorer(tau, g = identity, wtau = 1)
}
\arguments{
\item{tau}{Vector of quantile levels (in (0,1)) to score one.}

\item{g}{Vectorized transformation function of the response.}

\item{wtau}{Vector of weights to apply corresponding to each \code{tau}.}
}
\value{
A list of the following entries:

\itemize{
   \item \code{$tau} The inputted \code{tau} vector.
   \item \code{$g} The inputted \code{g} function.
   \item \code{$wtau} Vector of length matching \code{tau} of weights.
}
}
\description{
Indicate which scoring rule you'd like to use. Note that the scoring rule
is defined by the quantile levels, the transformation function, and the
across-quantile weight function.
}
\note{
This function isn't especially glamorous; it's just intended as a
simple way to "bundle" information together so that you don't have to work
with as many "moving parts" in an analysis. It does, however, provide
a convenient framework for dealing with different ways of specifying
weights on \code{tau}.
}
\examples{
tau <- space_taus(10)
scorer(tau)
scorer(tau, g=log, wtau=sqrt)
scorer(tau, wtau=1:10)
scorer(tau, wtau=2)
}
\seealso{
\code{\link{scoreq}} for evaluating the scoring rule.
}
