% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreq.R
\name{scoreq}
\alias{scoreq}
\alias{score_eval}
\title{Score Individual Quantiles}
\usage{
scoreq(y, yhat, tau, w = 1, g = identity, cmb = TRUE, wtau = NULL,
  na_omit = TRUE)

score_eval(y, yhat, sc, w = 1, cmb = TRUE, na_omit = TRUE, se = FALSE)
}
\arguments{
\item{y}{Vector of realizations}

\item{yhat}{Forecast matrix. Each row should be a forecast (of quantiles)
with corresponding realization in \code{y}; columns represent the level/index
of the quantile (corresponding to \code{tau}).}

\item{tau}{Vector of quantile indices of the quantiles.}

\item{w}{Vector of weights corresponding to the observations in
\code{yhat}. No need to normalize them. See details to see exactly how these
weights are used.}

\item{g}{Increasing univariate vectorized function to transform \code{y}
and \code{yhat} with (see computation details in the "details" section).
Default is the identity.}

\item{cmb}{Logical; should the scores be combined (via average)? \code{TRUE}
if so, \code{FALSE} to output a score matrix for each observation (rows)
and each quantile level (column).}

\item{wtau}{Function that accepts a vector of quantile indices and returns
an equally lengthed vector of weights to multiply the corresponding
individual quantile scores by. \code{NULL} for equal weights.}

\item{na_omit}{Logical; should observations leading to an \code{NA} score
(for any \code{tau}) be removed? \code{TRUE} by default. Warning message
appears when observations are removed.}

\item{sc}{The scoring rule to use, as in the output of the function
\code{\link{scorer}}.}

\item{se}{Logical; should an estimate of the standard error of the mean
estimate be returned as well? \code{TRUE} if so (which also overrides
the \code{cmb} argument, which is taken to be \code{TRUE}).}
}
\value{
If \code{se} is \code{TRUE}, returns a named vector of length two of the
average score (weighted by argument \code{w}) and standard error of the
average. The standard error is estimated by assuming iid scores, and is the
standard deviation of the scores times the root sum of squares of the
normalized weights \code{w}.

Here's what is output if \code{se} is \code{FALSE} (always the case with
the deprecated \code{scoreq} function).
If \code{cmb} is \code{FALSE}, returns the score matrix (see details for
how each score is computed) (rows correspond
to observations, and columns correspond to quantile indices \code{tau}).
Otherwise, a single numeric score is combined that is the average of the
score matrix.
}
\description{
Score quantile forecasts against realizations.
}
\details{
\code{scoreq} is deprecated. It doesn't allow for the computation of standard
error, and if asked to return the matrix of scores, it would incorporate
the across-observation weights into the scores, whereas \code{score_eval}
does not.

Here's how the score for the \eqn{i}'th observation
and the \eqn{k}'th quantile forecast for that observation is computed:
\deqn{wtau(\tau_k) (\tau_k - I(y<yhat_{ik}))(g(y) - g(yhat_{ik})),}
which is a proper scoring rule as shown in Gneiting and Raftery (2007).

To get a score for a particular observation, the average (not the sum)
is taken for each row. The scores aren't summed, so that the score doesn't
tend to infinity as we include more and more quantiles. Also, the
across-quantile weights, determined by the function \deqn{wtau}, are not
normalized, so that the individual scores don't tend to 0 as more
quantiles are included.
}
\note{
You could consider having the transformation function \code{g} transform
each observation differentially, by forcing it to accept a vector of
length equal to your data. This is useful to add seasonal trends, for
example.
}
\examples{
n <- 10
set.seed(364)
y <- rnorm(n)
tau <- c(0.8, 0.9)
sc <- scorer(tau)
yhat <- matrix(qnorm(tau), nrow=n, ncol=length(tau), byrow=TRUE)

score_eval(y, yhat, sc, cmb=FALSE)
score_eval(y, yhat, sc, se=TRUE)
score_eval(y, yhat, sc)
}
\references{
\itemize{
     \item{Gneiting, T. and Raftery, A. E. (2007). Strictly proper scoring
     rules, prediction, and estimation. Journal of the American Statistical
     Association, 102(477):359–378.}
}
}
