% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_taus.R
\name{space_taus}
\alias{space_taus}
\title{Get equally spaced quantile indices.}
\usage{
space_taus(K, tau_c = 0.9)
}
\arguments{
\item{K}{Integer. Number of quantile indices to obtain.}

\item{tau_c}{Value in the open interval (0,1) to take as the lower
quantile index bound.}
}
\value{
A vector of length \code{K} with the quantile indices.
}
\description{
Obtain \code{K} values in the middle of \code{K}
intervals of equal length within \code{(tau_c, 1)}.
}
