% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xylink.R
\name{xylink}
\alias{xylink}
\title{Extract information from a vine linking predictors and response}
\usage{
xylink(obj)
}
\arguments{
\item{obj}{Object of type \code{'rvine'}.}
}
\value{
A list of the following items:

\itemize{
     \item \code{$yvar}: Integer of the variable number of the response.
     \item \code{$xord}: Integer vector of the predictors listed below
     the response in the vine array (in order from upper to lower).
     \item \code{$p}: Integer; the number of predictors listed below the
     response.
     \item \code{$cops}: Character vector of the copula names linking the
     predictors and the response, in the order of \code{xord}.
     \item \code{$cpars}: List of numeric vectors, each vector being the
     copula parameters corresponding to the copulas in \code{cops}.
}
}
\description{
Given a vine, this function will assume the first variable in the last
column is a response variable, and the variables below it are predictors.
This function will return the variable numbers of the predictors and
response, as well as the copulas linking them.
}
\examples{
data(egdat)

## Work with this vine:
summary(egvine)

## Extract the linkage info:
xylink(egvine)

## What if we append another variable, with no links?
rv <- augment(egvine, 6)
xylink(rv)
}
