% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qsmooth.R
\name{get_qsmooth}
\alias{get_qsmooth}
\title{Compute Smoothed Quantile Data Frame}
\usage{
get_qsmooth(data, ycol = 1, xcol = 2, npoints = 100, h = 0.5,
  tau = space_taus(10))
}
\arguments{
\item{data}{Data frame containing the paired data}

\item{ycol, xcol}{Column numbers of the y and x variables.}

\item{npoints}{Number of points along the range of x to estimate quantiles at.}

\item{h}{bandwidth, in terms of number of standard deviations}

\item{tau}{Vector of quantile indices to regress over.}
}
\value{
Data frame with columns "x" (the grid of x values), "tau"
(quantile levels), and "y" (quantile estimate).
}
\description{
For x-y data, uses a moving window along x to estimate quantiles using a GPD
tail, using \code{\link{gpdqr}}.
}
\examples{
dat <- data.frame(y=rnorm(1000), x=runif(1000))
res <- get_qsmooth(dat)

head(dat)
library(ggplot2)
ggplot(dat, aes(x, y)) +
  geom_point() +
  geom_line(data=res, mapping=aes(x,y,group=tau), colour="red")
}
